pyproject.toml
[project]
name = "panoptes"
version = "0.1.0"
description = "Panoptes — vLLM ecosystem project dashboard"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "requests",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.10",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.12",
    "types-requests",
    "pre-commit>=4.3.0",
    # AES-GCM + PBKDF2 for the admin-only roster encryption CLI
    # (scripts/vllm/encrypt_roster.py). Must match docs/assets/js/token-vault.js.
    "cryptography>=42",
]

[build-system]
requires = ["setuptools>=69.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["scripts"]
exclude = ["tests*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Focused, practical rule set.  Style-churn rules (UP, RUF, SIM, N, C4)
# are deliberately off so the linter flags real bugs, not cosmetic drift.
select = ["E", "F", "W", "I", "B"]
ignore = [
    "E501",  # line-too-long — handled by formatter
    "E741",  # ambiguous variable name (l, I, O) — accepted in math/CI code
    "E701",  # multiple statements on one line (colon)
    "E702",  # multiple statements on one line (semicolon)
    "E402",  # module import not at top — occasionally needed for sys.path shims
    "F401",  # unused imports — tolerated; pruned opportunistically
    "F541",  # f-string without placeholders
    "F841",  # unused local variable (grandfathered — caught by mypy strict-mode later)
    "W293",  # whitespace on blank line
    "B007",  # unused loop control variable
    "B023",  # function uses loop variable (checker flags safe idioms)
    "B905",  # zip() without explicit strict=
    "I001",  # unsorted imports — cosmetic; can be fixed with `ruff check --fix`
]

[tool.ruff.lint.per-file-ignores]
# Tests can use asserts, long lines, and magic values freely.
"tests/**" = ["B011"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false       # legacy collectors return Any in many spots
warn_unused_configs = true
ignore_missing_imports = true
# Only type-check the refactored CI package.  Legacy collectors are
# excluded until they are modernised.
files = ["scripts/vllm/ci"]
disable_error_code = [
    "var-annotated",   # missing annotation on plain dict/list literals
    "arg-type",        # generic .sort()/setdefault signatures clash with dict[str, object]
    "return-value",    # narrowing unions when the function guarantees the type
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short --strict-markers"
